+++ modules/mysql/src/test/java/org/testcontainers/jdbc/JDBCDriverTest.java
@@ -11,7 +11,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 
-import static org.testpackage.VisibleAssertions.assertEquals;
+import static org.rnorth.visibleassertions.VisibleAssertions.assertEquals;
 
 /**
  *